+++ app/app.js
@@ -111,7 +111,7 @@ define([
             navigator.serviceWorker.register('sw.js').then(reg => {
                 setInterval(() => {
                     reg.update();
-                }, 100000);
+                }, 90000);
             });
         }
     };
@@ -224,7 +224,7 @@ define([
             $input.get(0).select();
             $input.get(0).setSelectionRange(0, 99999);
             document.execCommand('copy');
-            ALERT.show(GLOBAL.getLangJson().COPY_TO_CLIPBOARD, 'success');
+            ALERT.show(GLOBAL.getLangJson().COPIED_TO_CLIPBOARD, 'success');
             if (window.getSelection) {
                 if (window.getSelection().empty) { 
                     // Chrome