+++ app/features/chatbox/chatboxContent.js
@@ -271,6 +271,11 @@ define([
 
             if (getRoomById(roomInfo.id) && isInit) {
                 onGetMessageFromCache(roomInfo);
+
+                // update chat last read time
+                API.post(`chats/${roomInfo.id}/read`).then(() => {})
+                    .catch(err => console.error(err));
+
                 return;
             }
 
@@ -329,14 +334,15 @@ define([
             }
         },
 
-        onSyncRemove: (id) => {
+        onSyncRemove: (message) => {
+            const id = message.id.messageId;
             const $message = $(`[${ATTRIBUTE_MESSAGE_ID}="${id}"]`);
-            const $prevMessage = $message.prev();
-            const $prevMessageTwo = $prevMessage.prev();
-
-            $prevMessage.hasClass('not-mess-li') && $prevMessage.remove();
-            $prevMessageTwo.hasClass('not-mess-li') && $prevMessageTwo.remove();
-            $message.remove();
+            $message.find('.--mess').addClass('--message-removed').html(decodeStringBase64(message.message));
+            $message.find('.--mess').removeClass('fwme');
+            $message.find('.above-of-mess').removeClass('fwme');
+            $message.find('.btn-message-settings').hide();
+            $message.find('.--edited').addClass('hidden');
+            $message.find('.--double-check').addClass('hidden');
         },
 
         onSyncUpdate: (message) => {
@@ -344,6 +350,7 @@ define([
             const $message = $(`[${ATTRIBUTE_MESSAGE_ID}="${id}"]`);
 
             $message.find('.--mess').html(transformLinkTextToHTML(htmlEncode(decodeStringBase64(message.message))));       
+            $message.find('.--edited').removeClass('hidden');
         },
 
         onSearch: (searchMessageList, search) => {